Refactor custom tkinter classes to new location
