+++ core/ui/tk_var/__init__.py
@@ -1,0 +1,7 @@
+from .dict_var import DictVar
+from .list_var import ListVar
+
+__all__ = [
+    "DictVar",
+    "ListVar"
+]
