+++ core/ui/widgets/ro_text_w_var.py
@@ -1,0 +1,33 @@
+import tkinter as tk
+from tkinter import ttk
+
+class ReadOnlyTextWithVar(tk.Frame):
+    def __init__(self, parent, textvariable=None, **kwargs):
+        super().__init__(parent)
+        kwargs.pop("state", None) # Prevent double calling of state in Text Widget
+        self.textvariable = textvariable
+
+        self.text_widget = tk.Text(self, **kwargs, state="disabled")
+        self.scrollbar = ttk.Scrollbar(self, command=self.text_widget.yview)
+        self.text_widget.configure(yscrollcommand=self.scrollbar.set)
+
+        self.text_widget.grid(row=0, column=0, sticky="nesw")
+        self.scrollbar.grid(row=0, column=1, sticky="ns")
+        self.grid_columnconfigure(0, weight=1)
+        self.grid_rowconfigure(0, weight=1)
+
+        # Initialize text from textvariable
+        if self.textvariable is not None:
+            self.update_text_widget()
+            self.textvariable.trace_add("write", lambda *_: self.update_text_widget())
+
+    def update_text_widget(self):
+        """Update the Text widget when the StringVar changes."""
+        if self.textvariable is not None:
+            new_text = self.textvariable.get()
+            
+            # Enable text widget temporarily to modify text
+            self.text_widget.config(state="normal")
+            self.text_widget.delete("1.0", tk.END)
+            self.text_widget.insert("1.0", new_text)
+            self.text_widget.config(state="disabled")  # Re-disable input
